.Refactor moved buildinfo into sokar
